Adding configuration files for wx 2.8.x
These are configuration files for wxWidget 2.8.x used for building the official binary, they are not necessary for compilation but when used they make it possible to create a smaller binary with no dependencies.
These files are Windows only as wx on linux is different in this regard.
setup.h should be placed in %WXWIN%\include\wx\msw
config.vc should be placed in %WXWIN%\build\msw